+++ src/config/ci.php
@@ -26,7 +26,13 @@ return [
      * Regex to match file names and li
      *
      */
-    'regex_file_matcher' => '/([A-Za-z0-9\/._-]+):([1-9][0-9]*)/',
+    'regex_file_matcher' => '/([A-Za-z0-9\/._-]+)(?::| on line )([1-9][0-9]*)/',
+
+    /**
+     * Config file
+     */
+
+    'config_file' => config_path('ci.php'),
 
     /*
      * Regex to match file names and li